Add an extra ergonomic method for n2c chainsync
When actually consuming the chainsync client, initially the consumer
should call request_next repeatedly. However, once you are caught up,
you will get an "Await" response, which says we've requested the next
block, but the server doesn't have one for us, and will let us know when
one is available. If we request_next again, it fails.

In fact, even the example was misleading in this way!

The correct way to consume it was to check if you have agency, and if
so, either await or recieve.

If you tried to do this, though, especially when mixing it with other
future things (like select! on a cancellation token), then things could
get messy, because even though they return the same type, the impl
Future instances were opaque to the compiler.

This convenience method makes it nice and simple to idiomatically just
consume the next block as appropriate, and plays well with select!